move deprecated code to examples
